Sorted a couple of simple bugs
switched the instance cache VAO generation from updateForest to paintGL,
which seemed to sort out a load of problems. Also removed the final (as
far as I can see) reference to m_vertices in the body of createGeometry
and replaced it with the pointer vertices. Now generates something akin
to an actual forest, but some of the positioning is off.