Ok, I think I worked out what was happening:
basically, if I got this right, the problem was that I'd put a load of
the noise headers in a folder called 'noise' in 'include'. To use the
library, we also need to add Libnoise/include to the includePath, which
similarly contains a folder called 'noise'. I reckon the library needs
to access stuff from the noise folder, but when I put '$PWD/include'
first in the include path it tries to access from '$PWD/include/noise'
rather than 'Libnoise/include/noise' - I think the contents of these two
folders should actually be identical, but possibly the library also
needs to access other files relative to 'Libnoise/include/noise' and is
instead searching relative to '$PWD/include/noise', or something like
that. Hence the problem.

Anyhow, I solved it by just deleting the noise directory in include, and
they are just accessed from the 'Libnoise/include' on my laptop instead
(need to check that I wasn't relying on these headers in the linux
setup). I kept noiseutils.h in a new folder noiseutils though since this
file does need to be accessed